pyproject.toml
[tool.poetry]
name = "simpleapi"
version = "0.1.0"
description = ""
authors = ["azamtoiri <azamtoiri@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django-rest-framework = "^0.1.0"
django = "^5.1.3"
pillow = "^11.0.0"
django-cors-headers = "^4.6.0"
django-filter = "^24.3"
jsonschema = "^4.23.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
